requirements.txt
Flask
azure-ai-formrecognizer
flask-cors
